🔥 ✨ 💄 (UI): Major changes to folder structure
