+++ client/src/components/ExploreCard.tsx
