--- client/src/components/PageNotFound/PageNotFound.tsx
@@ -1,11 +1,0 @@
-// import React from 'react';
-
-function PageNotFound() {
-	return (
-		<div>
-			<h1>404 PAGE NOT FOUND</h1>
-		</div>
-	);
-}
-
-export default PageNotFound;
