+++ client/src/components/PublishedText.tsx
