+++ client/src/components/SharedCard.tsx
