+++ client/src/components/styles/Logo.css
