+++ client/src/components/styles/Splash.css
