+++ client/src/context/AuthProvider.tsx
@@ -1,19 +1,15 @@
-import { createContext, ReactNode, useState } from "react";
-
-type Props = {
-	children: ReactNode;
-};
+import { createContext, useState } from "react";
 
 const AuthContext = createContext({} as any);
 
-export const AuthProvider = ({ children }: Props) => {
-	const [isAuthenticated, setIsAuthenticated] = useState({});
+export const AuthProvider = ({ children }: any) => {
+    const [auth, setAuth] = useState({});
 
-	return (
-		<AuthContext.Provider value={{ isAuthenticated, setIsAuthenticated }}>
-			{children}
-		</AuthContext.Provider>
-	);
+    return (
+        <AuthContext.Provider value={{ auth, setAuth }}>
+            {children}
+        </AuthContext.Provider>
+    );
 };
 
 export default AuthContext;
