+++ client/src/hooks/useAuth.tsx
@@ -1,7 +1,6 @@
 import { useContext } from "react";
 import AuthContext from "../context/AuthProvider";
 
-
 function useAuth() {
     return useContext(AuthContext);
 }
